+++ frontendweb/controllers/controller.php
@@ -1,0 +1,16 @@
+<?php
+class MvcController{
+    public function plantilla(){
+        include "views/template.php";
+    }
+    public function enlacesPaginasController(){
+        if(isset($_GET["action"])){
+            $enlacesController = $_GET ["action"];
+        }else{
+            $enlacesController = "home.php";
+        }
+        $respuesta = EnlacesPaginas ::enlacesPaginasModel($enlacesController);
+        include $respuesta;
+    }
+}
+?>
