+++ frontendweb/css/styles.css
@@ -14,7 +14,6 @@
 
 .headerStyle {
     background-color: white;
-    color: black;
     top: 0px;
     width: 100%;
     height: fit-content;
@@ -27,8 +26,8 @@
 }
 
 a {
+    color: white;
     text-decoration: none;
-    color: #D9D9D9;
 }
 
 .indexStyle {
@@ -43,28 +42,4 @@ a {
 .divTituloLogin {
     text-align: center;
     font-weight: bold;
-}
-.navTemplate{
-    color: black;
-}
-.socialnetwork{
-    height: 25px;
-}
-.navBg{
-    color: black;
-    background-color: lightgray;
-}
-.navbar-brand {
-    display: flex;
-    align-items: center;
-    
-}
-
-.custom-nav .navbar-nav li:not(:last-child) a {
-    margin-right: 50px;
-}
-
-.custom-nav {
-    display: flex;
-    align-items: center;
 }
