+++ frontendweb/redireccion.php
@@ -1,10 +1,8 @@
 <?php
 session_start();
-require_once "controllers/controlleroficinista.php";
-require_once "servicios/modeloficinista.php";
+require_once "controllers/controller.php";
+require_once "servicios/model.php";
 
 $mvc = new MvcController();
-
-$mvc -> plantillaoficinista(); 
-
+$mvc -> plantilla(); 
 ?>
