+++ frontendweb/servicios/model.php
@@ -1,0 +1,11 @@
+<?php
+class EnlacesPaginas
+{
+    public static function enlacesPaginasModel($enlacesModel)
+    {
+
+        $module = "views/modules/home.php";
+        return $module;
+    }
+}
+?>
