--- frontendweb/servicios/modeloficinista.php
@@ -1,17 +1,0 @@
-<?php
-class EnlacesPaginas
-{
-    public static function enlacesPaginasModelOficinista($enlacesModel)
-    {
-        if($enlacesModel == "buses" ||
-            $enlacesModel == "frequencies" ||
-            $enlacesModel == "sales" ||
-            $enlacesModel == "validation"){
-                $module = "views/modules/".$enlacesModel."oficinista.php";
-            }else{
-            $module = "views/modules/homeoficinista.php";
-        }
-        return $module;
-    }
-}
-?>
