+++ frontendweb/views/modules/home.php
@@ -1,0 +1,2 @@
+<link rel="stylesheet" href="css/styles.css">
+<img src="img/loginbg.jpg" alt="" class="bgHome">
