--- frontendweb/views/modules/homeoficinista.php
@@ -1,7 +1,0 @@
-<link rel="stylesheet" href="css/styles.css">
-<div class="indexStyle">
-    <br><br><br><br>
-    <p>Welcome to <img class="avatar" src="img/safey.jpg"></p>
-
-    <p>With safety you can travel safely to all your favorite destinations!</p>
-</div>
